wip: use redis to cache posts

+++ backend/scripts/restart-chain-docker.sh
@@ -1,27 +1,38 @@
 #!/bin/bash
 
 # Stop and remove containers, networks
-echo "Stopping and removing containers, networks..."
+echo "Stopping and optionally removing containers, networks..."
 docker compose down
+    cat << EOI
 
-# Remove specified volumes to remove all state and start fresh
-echo "Removing specified volumes..."
-docker volume rm backend_redis_data
-docker volume rm backend_chainstorage
-docker volume rm backend_ipfs_data
-docker volume rm backend_node_cache
+┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
+┃ Do you want to remove the docker volumes [y/n]:                                             ┃
+┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
 
-# Start specific services in detached mode
-echo "Starting frequency services..."
-docker compose up -d frequency
+EOI
+    read REMOVE_VOLUMES
 
-# Wait for 15 seconds
-echo "Waiting 15 seconds for Frequency to be ready..."
-sleep 15
+    if [[ $REMOVE_VOLUMES =~ ^[Yy]$ ]]
+    then
+        # Remove specified volumes to remove all state and start fresh
+        echo "Removing specified volumes..."
+        docker volume rm backend_redis_data
+        docker volume rm backend_chainstorage
+        docker volume rm backend_ipfs_data
+        docker volume rm backend_node_cache
 
-# Run npm run local:init 
-echo "Running npm run local:init to provision Provider with capacity, etc..."
-npm run local:init
+        # Start specific services in detached mode
+        echo "Starting frequency services..."
+        docker compose up -d frequency
+
+        # Wait for 15 seconds
+        echo "Waiting 15 seconds for Frequency to be ready..."
+        sleep 15
+
+        # Run npm run local:init 
+        echo "Running npm run local:init to provision Provider with capacity, etc..."
+        npm run local:init
+    fi
 
 # Start all services in detached mode
 echo "Starting all services..."
